+++ sdk/resourcemanager/recoveryservices/armrecoveryservices/models.go
@@ -179,6 +179,11 @@ type IdentityData struct {
 	TenantID *string `json:"tenantId,omitempty" azure:"ro"`
 }
 
+// ImmutabilitySettings - Immutability Settings of vault
+type ImmutabilitySettings struct {
+	State *ImmutabilityState `json:"state,omitempty"`
+}
+
 // JobsSummary - Summary of the replication job data for this vault.
 type JobsSummary struct {
 	// Count of failed jobs.
@@ -660,6 +665,12 @@ type SKU struct {
 	Tier *string `json:"tier,omitempty"`
 }
 
+// SecuritySettings - Security Settings of the vault
+type SecuritySettings struct {
+	// Immutability Settings of a vault
+	ImmutabilitySettings *ImmutabilitySettings `json:"immutabilitySettings,omitempty"`
+}
+
 // SystemData - Metadata pertaining to creation and last modification of the resource.
 type SystemData struct {
 	// The timestamp of resource creation (UTC).
@@ -870,6 +881,9 @@ type VaultProperties struct {
 
 	// The redundancy Settings of a Vault
 	RedundancySettings *VaultPropertiesRedundancySettings `json:"redundancySettings,omitempty"`
+
+	// Security Settings of the vault
+	SecuritySettings *SecuritySettings `json:"securitySettings,omitempty"`
 
 	// Details for upgrading vault.
 	UpgradeDetails *UpgradeDetails `json:"upgradeDetails,omitempty"`